03-aidd/pyproject.toml
[project]
name = "03-aidd"
version = "0.1.0"
description = "Telegram-бот с LLM-ассистентом"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


